+++ cli/src/CHANGELOG.md
@@ -1,3 +1,6 @@
 # Changelog
 
+- 2025-10-28 18:19:00 — Remove MongoDB integration and update default model ([details](docs/changes/2025-10-28_18-19-00.md))
+
+
 - 2025-10-25 05:45:40 — Release Documentation ([details](docs/changes/2025-10-25_05-45-40.md))